+++ bill-of-material/gradle.properties
@@ -1,0 +1,6 @@
+ARTIFACT_NAME=android-bom
+NAME=MoEngage Android Dependency BOM
+POM_DESCRIPTION=Bill Of Materials for MoEngage's Android SDK
+RELEASE_VARIANT=javaPlatform
+JAVA_PLATFORM=true
+VERSION_NAME=0.0.1
