+++ catalog/gradle.properties
@@ -1,0 +1,6 @@
+ARTIFACT_NAME=android-dependency-catalog
+NAME=MoEngage Android Dependency Catalog
+POM_DESCRIPTION=Dependency Catalog for MoEngage's Android SDK
+RELEASE_VARIANT=versionCatalog
+VERSION_CATALOG=true
+VERSION_NAME=6.0.1
