+++ gradle.properties
@@ -1,10 +1,7 @@
 #Thu Sep 25 13:03:09 UTC 2025
-ARTIFACT_NAME=android-dependency-catalog
 GROUP=com.moengage
 HOST=CENTRAL_PORTAL
 IN_MEMORY_SIGNING=true
-NAME=MoEngage Android Dependency Catalog
-POM_DESCRIPTION=Dependency Catalog for MoEngage's Android SDK
 POM_DEVELOPER_EMAIL=sdksubscriptions@moengage.com
 POM_DEVELOPER_ID=moengage
 POM_DEVELOPER_NAME=MoEngage
@@ -14,5 +11,3 @@ POM_SCM_CONNECTION=scm\:git@github.com\:moengage/android-dependency-catalog.git
 POM_SCM_DEV_CONNECTION=scm\:git@github.com\:moengage/android-dependency-catalog.git
 POM_SCM_URL=https\://moengage.com
 POM_URL=https\://moengage.com
-RELEASE_VARIANT=versionCatalog
-VERSION_NAME=6.0.1
