+++ settings.gradle.kts
@@ -6,4 +6,17 @@ pluginManagement {
     }
 }
 
+enableFeaturePreview("TYPESAFE_PROJECT_ACCESSORS")
+dependencyResolutionManagement {
+    repositoriesMode.set(RepositoriesMode.FAIL_ON_PROJECT_REPOS)
+    repositories {
+        google()
+        mavenCentral()
+    }
+}
+
+include(
+    ":bill-of-material",
+    ":catalog"
+)
 rootProject.name = "android-dependency-catalog"
